+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     + \#52300: count newlines as two characters in character counter
     + \#52889: Resolve invalid url_regex Regex (reported by Judge's JS)
     + Make sure filtering selects don't overlap submit button
+    + \#52646: Fix cloning issue when unique attributes also have a length validation
 
 ## 1.2